refactor: naive_codec.py -> pickle_codec.py
